// src/enums/eBrdUtsMagnitude.ts
/** ---------------------------------------------------------------------------
 * @module [Brd/Uts]
 * @author APG
 * @version 0.2 APG 20230418
 * ----------------------------------------------------------------------------
 */

/**
 * Magnitude of values used to convert them to units
 */
export enum eBrdUtsMagnitude {
  unity = 1,
  mega = 0.000001,
  pair = 0.5,
  deca = 0.1,
  etto = 0.01,
  kilo = 0.001,
  deci = 10,
  centi = 100,
  milli = 1000,
  nano = 1000000
}